requirements.txt
langchain==0.2.5
langchain_community==0.2.5
langchain_groq==0.1.5
faiss-cpu==1.8.0 
langchain-chroma==0.1.4
langchain_huggingface==0.0.3
python-dotenv==1.0.1
streamlit==1.35.0
pypdf==4.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
